docker/requirements.txt
requests==2.32.3
kafka-python==2.0.2
websockets==12.0
pycryptodome==3.20.0
urllib3==2.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
brotli==1.1.0
pymongo==4.7.0
pyspark==3.5.1
pandas
s3fs
